+++ config/initializers/passpartu.rb
@@ -1,7 +1,10 @@
+# frozen_string_literal: true
+
 # this is example of initializer
 # this file is required for gem's test cases
 
 Passpartu.configure do |config|
   # config.policy_file = './config/passpartu.yml'
   # config.raise_policy_missed_error = true
+  # config.check_waterfall = false
 end