+++ lib/passpartu.rb
@@ -1,13 +1,18 @@
+# frozen_string_literal: true
+
 require 'passpartu/version'
 require 'yaml'
 require_relative 'passpartu/patcher'
 require_relative 'passpartu/verify'
 require_relative 'passpartu/block_verify'
 require_relative 'passpartu/validate_result'
+require_relative 'passpartu/check_waterfall'
 require_relative 'passpartu/user' # for testing only
 
 module Passpartu
   class Error < StandardError; end
+  class PolicyYmlNotFoundError < StandardError; end
+  class WaterfallError < StandardError; end
 
   def self.included(policy_class)
     Passpartu::Patcher.call(policy_class)
@@ -27,17 +32,23 @@ module Passpartu
   end
 
   class Config
-    attr_accessor :policy, :raise_policy_missed_error
+    attr_accessor :policy, :raise_policy_missed_error, :check_waterfall
     attr_reader :policy_file
 
+    DEFAULT_POLICY_FILE = './config/passpartu.yml'
+
     def initialize
-      @policy_file = './config/passpartu.yml'
-      @policy = YAML.load_file(policy_file)
+      @policy_file = DEFAULT_POLICY_FILE
+      @policy = YAML.load_file(policy_file) if File.exist?(policy_file)
       @raise_policy_missed_error = true
+      @check_waterfall = false
     end
 
     def policy_file=(file = nil)
-      @policy_file = file || './config/passpartu.yml'
+      @policy_file = file || DEFAULT_POLICY_FILE
+
+      raise PolicyYmlNotFoundError unless File.exist?(policy_file)
+
       @policy = YAML.load_file(policy_file)
     end
   end