+++ CHANGELOG.md
@@ -11,7 +11,7 @@ To enable feature change waterfall_rules to `true` in config/initializers/passpa
     config.waterfall_rules = true
   end
 ```
-If policy rule is missed, but the key above value equals true - positive policy (true)
+If policy rule is missed, but the key above value equals true - positive policy (true).
 It work with any number of keys!
 ```yml
   admin: