+++ CHANGELOG.md
@@ -1,6 +1,38 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.0.0] - 2020-04-05
+### Added
+#### Waterfall check
+Allow or restrict absolutely everything for particular role or/and particular domain.
+```ruby
+# ./config/initializers/passpartu.rb
+
+Passpartu.configure do |config|
+  config.check_waterfall = true
+end
+```
+
+```yml
+# ./config/passpartu.yml
+
+super_admin: true
+super_looser: false
+medium_looser:
+  orders:
+    create: true
+    delete: false
+  products: true
+```
+```ruby
+user_super_admin.can?(:do, :whatever, :want) # true
+user_super_loser.can?(:do, :whatever, :want) # false
+user_medium_loser.can?(:orders, :create) # true
+user_medium_loser.can?(:orders, :delete) # false
+user_medium_loser.can?(:products, :create) # true
+user_medium_loser.can?(:products, :create, :and_delete) # true
+```
+
 ## [0.6.0] - 2019-04-01
 ### Added
 - Only roles attribute to verifier